RM#32972 Refactor code in a new service
